passit: optimise readUint32n for n=1
If n is 1, the result of readUint32n will always be 0. In that case we
can optimise by avoiding the read from r and immediately return 0.

It has particular advantage in the case of the regexp generator as now
fixed repeats like x{3} will no longer cause a read to occur. That
also means x{3} and xxx will now be identical.

This does impact r and the output, but doesn't it impact the amount of
entropy / randomness in the output.