passit: use AES-CTR instead of math/rand in tests
math/rand is a really poor choice for generating passwords and using it
in tests might imply to someone that it's suitable for low-value
passwords - this is absolutely not true. Instead we use a deterministic
CSPRNG based on AES-CTR. This is a safe choice as long as it is
suitably keyed.

We use AES-CTR rather than ChaCha20 or other stream ciphers because it's
included in the standard library and we don't need to import x/crypto.