passit: enable pruning of unused embedded strings
Writing a test for this is possible (see go.dev/cl/402555) but
non-trivial. In this case I've tested this manually and confirmed that
importing go.tmthrgd.dev/passit doesn't automatically pull in all the
embedded strings.

This is important as the four embedded strings currently consume
114.5 kB. A number that is likely to increase especially as future
Unicode versions add new emoji.