cmd/twoproblems: use case-insensitive name lookup for wordlist capture
When operating in case-insensitive mode, the special capture content is
converted to uppercase (the lowest rune from unicode.SimpleFold). This
causes the wordlist name lookup to fail:
  $ go run ./cmd/twoproblems '(?i:abc-(?P<word>eff)-xyz)'
  twoproblems: failed to parse "(?i:abc-(?P<word>eff)-xyz)" pattern: twoproblems: unsupported wordlist "EFF"
  exit status 1

We get around this by performing a case-insensitive comparison. The count
is unaffected, but there's no way to unmangle the separator. If this is
an issue, callers will need to disable case-insensitive mode for that
part of the pattern with (?-i:...).