passit: ensure that Z{0} and Z{1} do not read from r in RegexpParser
Z{0} should be equivalent to (?:) and Z{1} should be equivalent to Z.
This locks that behaviour in.

We're already testing other values like Z{2}.

We explicitly test .{0} to ensure that the inner . doesn't read
from r either. That makes sure we don't first evaluate the sub
expression before deciding whether to toss it or not.